refactor: fold source changes into updated main
