+++ .changeset/blue-cycles-give.md
@@ -1,0 +1,5 @@
+---
+"@adobe/aio-commerce-lib-app": patch
+---
+
+Improve error logging and fix an issue where the project root directory was not derived correctly in some situations.
