+++ .changeset/wild-geckos-knock.md
@@ -1,0 +1,5 @@
+---
+"@adobe/aio-commerce-lib-config": patch
+---
+
+Fix `encryption setup` failing when the target `.env` file does not exist, the file is now created if missing.
