+++ packages-private/scripting-utils/source/filesystem/helpers.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Adobe. All rights reserved.
+ * Copyright 2026 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -10,12 +10,16 @@
  * governing permissions and limitations under the License.
  */
 
-/** biome-ignore-all lint/performance/noBarrelFile: Public entrypoint for the filesystem helpers. */
+import { writeFile } from "node:fs/promises";
 
 /**
- * This module exports shared filesystem utilities for the AIO Commerce SDK.
- * @packageDocumentation
+ * Touch a file, creating it if it doesn't exist.
+ * @param filePath The path of the file to touch
  */
-
-export * from "./helpers";
-export * from "./temp";
+export async function touch(filePath: string) {
+  await writeFile(filePath, "", { flag: "wx" }).catch((e) => {
+    if (e.code !== "EEXIST") {
+      throw e;
+    }
+  });
+}
