+++ packages/aio-commerce-lib-app/source/commands/generate/actions/config.ts
@@ -1,5 +1,5 @@
 /*
- * Copyright 2025 Adobe. All rights reserved.
+ * Copyright 2026 Adobe. All rights reserved.
  * This file is licensed to you under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License. You may obtain a copy
  * of the License at http://www.apache.org/licenses/LICENSE-2.0
@@ -149,6 +149,7 @@ export function buildAppManagementExtConfig(
     "installation.customInstallationSteps",
     "eventing.commerce",
     "eventing.external",
+    "webhooks",
     "adminUiSdk",
   ];
