+++ packages/aio-commerce-lib-app/source/commands/utils.ts
@@ -10,10 +10,22 @@
  * governing permissions and limitations under the License.
  */
 
+import { join } from "node:path";
+
 import consola from "consola";
 import * as prettier from "prettier";
 
 import { parseCommerceAppConfig } from "#config/index";
+
+import {
+  APP_MANIFEST_FILE,
+  CONFIG_SCHEMA_FILE_NAME,
+  CONFIGURATION_EXTENSION_POINT_ID,
+  EXTENSIBILITY_EXTENSION_POINT_ID,
+  GENERATED_ACTIONS_PATH,
+  GENERATED_PATH,
+  getExtensionPointFolderPath,
+} from "./constants";
 
 /** Format file content using prettier, inferring the parser from the file path. */
 export function prettierFormat(content: string, filepath: string) {
@@ -38,4 +50,56 @@ export async function loadAppManifest() {
   consola.debug("Loaded app commerce config");
 
   return appConfig;
+}
+
+/**
+ * Path to an extension point's `.generated` directory, relative to the project root.
+ * @param extensionPointId - The extension point ID, e.g. "commerce/extensibility/1"
+ */
+export function getGeneratedDir(extensionPointId: string) {
+  return join(getExtensionPointFolderPath(extensionPointId), GENERATED_PATH);
+}
+
+/**
+ * Path to an extension point's generated actions directory, relative to the project root.
+ * @param extensionPointId - The extension point ID, e.g. "commerce/extensibility/1"
+ */
+export function getActionsDir(extensionPointId: string) {
+  return join(
+    getExtensionPointFolderPath(extensionPointId),
+    GENERATED_ACTIONS_PATH,
+  );
+}
+
+/**
+ * Path to a specific generated action file, relative to the project root.
+ * @param extensionPointId - The extension point ID, e.g. "commerce/extensibility/1"
+ * @param actionName - The name of the action, e.g. "my-action"
+ */
+export function getActionPath(extensionPointId: string, actionName: string) {
+  return join(getActionsDir(extensionPointId), `${actionName}.js`);
+}
+
+/**
+ * Path to an extension point's `ext.config.yaml`, relative to the project root.
+ * @param extensionPointId - The extension point ID, e.g. "commerce/extensibility/1"
+ */
+export function getExtConfigPath(extensionPointId: string) {
+  return join(getExtensionPointFolderPath(extensionPointId), "ext.config.yaml");
+}
+
+/** Path to the generated app manifest file, relative to the project root. */
+export function getManifestPath() {
+  return join(
+    getGeneratedDir(EXTENSIBILITY_EXTENSION_POINT_ID),
+    APP_MANIFEST_FILE,
+  );
+}
+
+/** Path to the generated configuration schema file, relative to the project root. */
+export function getSchemaPath() {
+  return join(
+    getGeneratedDir(CONFIGURATION_EXTENSION_POINT_ID),
+    CONFIG_SCHEMA_FILE_NAME,
+  );
 }
