+++ packages/aio-commerce-lib-config/source/commands/encryption/setup/main.ts
@@ -13,6 +13,7 @@
 import { dirname, join } from "node:path";
 
 import { replaceEnvVar } from "@aio-commerce-sdk/scripting-utils/env";
+import { touch } from "@aio-commerce-sdk/scripting-utils/filesystem/helpers";
 import { findNearestPackageJson } from "@aio-commerce-sdk/scripting-utils/project";
 import consola from "consola";
 
@@ -21,7 +22,8 @@ import { generateEncryptionKey } from "#utils/encryption";
 const ENCRYPTION_KEY_ENV_VAR = "AIO_COMMERCE_CONFIG_ENCRYPTION_KEY";
 
 /** Sets up an encryption key for password field at the given environment file path. */
-export function run(envPath: string) {
+export async function run(envPath: string) {
+  await touch(envPath);
   process.loadEnvFile(envPath);
 
   if (!process.env[ENCRYPTION_KEY_ENV_VAR]) {
@@ -42,7 +44,7 @@ export async function exec() {
     const dir = packageJson ? dirname(packageJson) : process.cwd();
     const envPath = join(dir, ".env");
 
-    run(envPath);
+    await run(envPath);
   } catch (error) {
     consola.error(error);
     process.exit(1);
